Some fixes
Here are some fixes to make the code better and more readable.

The first fix is the replacement of the navigator.userAgent command. It's replaced by a more reliable and safer option especially for future versions of browsers.

The second change is for the widget.js file. I replaced the createCustomDialog function. Beacuse it was too long and unreadable. I broke it to pieces with some specialized function, so that is more readable and more to the poit